requirements.txt
feedparser
google-genai
firebase-admin
trafilatura
geopy